Move DNS polyfills to a PHP extension and add missing constants (#27)
## Motivation for the change, related issues

The original implementation of the DNS polyfill didn't register `DNS_*`
constants which are [expected to exist in
PHP](https://www.php.net/manual/en/network.constants.php#constant.dns-any).
To fix this this PR registered the required DNS constants.

## Implementation details

The constants were already defined in the polyfill and now we register
them as PH constants using `REGISTER_LONG_CONSTANT`.

I didn't want to add more code to the `php_wasm.c` file that wasn't
directly related to PHP-wasm, so I moved the DNS polyfill code to an
extension.

## Testing Instructions (or ideally a Blueprint)

- ~~CI~~
- [unit tests currently
fail](https://systemsrequests.wordpress.com/2025/02/11/github-actions-runs-slowly-and-are-canceled-on-the-automattic-wordpress-playground-private-repository/)
so you should run tests locally
  - `nvm use 20` (to avoid crypto errors)
  - `npx nx run php-wasm-node:test`