Store WordPress site cookies in the browser instead of a custom Cookie Store (#20)
## Motivation for the change, related issues

A cookie store is needed on the web where custom Response objects may
not be assigned `Set-Cookie` headers. Otherwise, `Set-Cookie` headers
received in PHP responses are effectively discarded. But a cookie store
isn't needed for Playground CLI, and applying the same cookies to all
requests to a CLI-based server interferes with clients that want to use
other WP auth schemes.

This PR gives the browser-based PHP worker the responsibility for
maintaining a cookie store per Playground scope.

This feature was previously implemented by @brandonpayton in
[Automattic/wordpress-playground/pull/8](https://github.a8c.com/Automattic/wordpress-playground/pull/8).
We also explored using [document.cookie instead of the cookie
store](https://github.com/Automattic/wordpress-playground-private/pull/21#issuecomment-2665652952),
but we couldn't come up with a clean implementation due to browser
restrictions.

## Implementation details

This PR makes the cookie store a browser-based PHP worker concern so it
can manage which cookie store to apply to which requests.

Originally, I wanted to make the Service Worker responsible for the
synthetic cookie store because it is responsible for all request routing
and generally sits between the browser client and the PHP "server", but
it turned out to be quite awkward because:
- the Service Worker doesn't currently know when Playground site scopes
come and go, making cleanup difficult
- Service Workers can be terminated and restarted while client tabs are
still open. This meant that we'd have to persist cookies in something
like IndexedDB, and both service worker lifetime uncertainty and
ignorance about the lifetime of site scopes made it difficult to know
when old cookie stores could be cleared.

If the cookie store is kept with the PHP worker, then the cookies will
naturally come and go as Playground sites are loaded and unloaded.

## Testing Instructions (or ideally a Blueprint)

- CI