Remove Data Liberation PHP modules (#79)
Removes the PHP Data Liberation plumbing originally introduced in
https://github.com/WordPress/wordpress-playground/pull/1888. The
codebase has since moved to https://github.com/Automattic/php-toolkit
and is not maintained in the Playground repository anymore. It's only a
cause of test failures at the moment for little added benefit.

This PR also removes the possibility of importing WXR files using the
data liberation importer. It will have to be re-architected anyway to
accomodate for the changes in the php-toolkit repository structure.

 ## Testing instructions

This PR mostly removes code. Just confirm the CI checks pass