[Website] Drop the static <link rel="manifest"> tag and generate one in JavaScript instead (#78)
#67 introduced JavaScript logic to update the URL in the `<link
rel="manifest">` tag. This PR removes the server-rendered `<link>` tag
entirely and generates it in JavaScript instead.

I noticed Safari sometimes ignores the runtime updates to the URL of an
existing manifest tag. At the same time, it seems to respect the URL of
a freshly created manifest.

 ## Testing instructions

In Safari, Chrome, and Firefox:

Build Playground, run it via a PHP dev server on a https domain (I've
used ngrok), start it with a Blueprint and a custom `?app_name`
parameter, and confirm you can install it as a named PWA that runs that
Blueprint on boot.