Update Nx, Vite, and ESLint dependencies (#35)
## Motivation for the change, related issues

When I am learning about Nx or Vite or troubleshooting potential issues
with them, I find that our versions are old enough that I am not sure
whether docs and merged bug fixes apply the versions we're using.

This is a PR to explore updating and dispelling that uncertainty and
doubt.

**UPDATE (2025-04-16)** This PR is also required in order to use the
[eslint-plugin-erasable-syntax-only](https://www.npmjs.com/package/eslint-plugin-erasable-syntax-only)
package. That package depends on a newer ESLint version than we're using
today, and we cannot update that separately because our older version of
Nx depends on an older ESLint.

## Implementation details

- Use `nx migrate` to migrate to latest feasible version.
- Update Vite dependencies to latest versions compatible with the nx
packages.
- Fix any issues

## Testing Instructions (or ideally a Blueprint)

- CI