[PHP-wasm Node] Fix Identifier '__filename' has already been declared (#143)
## Motivation for the change, related issues

`__filename` and `__dirname` are available in CommonJS but unavailable
in ES modules. ([Node.js
docs](https://nodejs.org/api/esm.html#no-__filename-or-__dirname)).

In https://github.com/Automattic/wordpress-playground-private/pull/129
we made sure `__filename` and `__dirname` are available, but this broke
ES module support and the latest version of PHP-wasm (1.0.36) can't
initialize the PHP environment because of errors like `The "path"
argument must be of type string or an instance of URL. Received
undefined` and `SyntaxError: Identifier '__filename' has already been
declared`

## Implementation details

This PR addresses the issue by creating only the `__dirname` and
`__filename` constants only if they weren't previously defined. The
downside is that we need to define them as vars and not consts.

The PR also adds package install tests to CI, and the test creates
temporary packages from the current commit, installs them into an ESM
and CJS project,t and runs simple tests that boot WordPress using
`runCLI`.

## Testing Instructions (or ideally a Blueprint)

I tested this by manually updating PHP-wasm and Playground packages in
Studio and a personal project.
I also tested that we can run Playground CLI using `npx PACKAGE_URL
server`, and that we can use the CLI in Windows and MacOS using Bun.

- Run `npx nx debug playground-cli --php=8.3 server` and confirm it
still works

---------

Co-authored-by: Adam Zieliński <adam@adamziel.com>