[Website] Open external links in a new tab/window (#2468)
## Motivation for the change, related issues

Adds `target="_blank"` to external links in Playground web (not CLI).
This ensures external links are loaded in a new tab and not inside the
Playground iframe.

## Implementation details

Adds `a.target = '_blank';` via a `<script>` echoed by an mu-plugin.

## Testing Instructions (or ideally a Blueprint)

Go to wp-admin and click the `WordPress` link in this part of the
footer:

> Thank you for creating with
[WordPress](https://wordpress.org/).Version 6.8.2

cc @fellyph – While looking at your Blueprint I realized this is
something we should solve at the platform level.

cc @JanJakes for reviews