Fix corrupt zip download from Site Editor export (#2531)
## Motivation for the change, related issues

The purpose of this PR is to fix the corrupt zip download from the Site
Editor "Export" feature.

Fixes #2528

## Implementation details

This PR adds a `stdoutBytes` property to StreamedPHPResponse and updates
`PHPResponse.fromStreamedResponse()` to use that.

Prior to this change, `fromStreamedResponse()` used the `stdoutText`
property which took a byte stream that was decoded on the fly to text
and then re-encoded by the accessor to bytes. The bytes of a zip file
going through that transform were being corrupted. Without that
transform, the zip file bytes are not corrupted.

## Testing Instructions (or ideally a Blueprint)

- Automated: CI
- Manual
  - Run `npx nx unbuilt-asyncify playground-cli -- server`
  - Open the site in a browser
  - Navigate to the Site Editor
- Open the upper-right hamburger menu and click the "Export" menu item
to download a theme export zip.
- Confirm that the downloaded zip can be extracted and that the contents
do not look obviously broken.