Playground CLI: Support --auto-mount=path option (#2525)
## Motivation for the change, related issues

It is impossible (or at least awkward) to test the Playground CLI
`--auto-mount` option locally using the `npx nx unbuilt-jspi
playground-cli` family of commands. The reason is that `--auto-mount`
tries to auto mount the current working directory, but those commands
must be executed with the Playground source directory.

For this case, it would be useful to be able to pass a path to the
`--auto-mount` option.

In addition, it is more powerful and flexible for to regular users to be
able to pass an explicit `--auto-mount=path`.

## Implementation details

This PR:
- Changes the type of the `--auto-mount` option to a string.
- Updates the `--auto-mount` yargs description.
- Updates `runCLI()` to default auto-mount to the current working
directory if `--auto-mount` is specific but no path is provided.
- Adds tests passing an explicit path to auto-mount.

## Testing Instructions (or ideally a Blueprint)

- CI